requirements.txt
# FastAPI 和服務器
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP 和檔案處理
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0

# 配置和驗證
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具和日誌
python-dotenv==1.0.0
loguru==0.7.2

# PDF 生成
reportlab==4.0.7

# CORS 支援
python-multipart==0.0.6
